[WASM] WASM in worker
