+++ .babelrc
@@ -1,3 +1,9 @@
 {
-    "presets": ["env"]
+    "presets": ["env"],
+    "plugins": [
+      ["transform-runtime", {
+        "polyfill": true,
+        "regenerator": true
+      }]
+    ]
 }
